refactor(serializers): modularize serializers.py into domain-based modules
Extract 23 serializers from monolithic serializers.py (922 lines) into
8 domain-specific files for better maintainability and organization.

Structure:
- serializers/__init__.py: Re-exports for backwards compatibility
- serializers/usuario.py: UserSlim, UsuarioOption, UsuarioAdmin, Group (4)
- serializers/organizacao.py: Municipio*, Projeto*, Gerencia, TipoEvento, Produto (7)
- serializers/solicitacao.py: Solicitacao, Participation, EventDetail, AuditLogTimeline (4)
- serializers/agenda.py: AvailabilityBlock (1)
- serializers/compra.py: Compra (1)
- serializers/workflow.py: AcaoControle, AcaoDAT*, Deslocamento (4)
- serializers/auditoria.py: AuditLog (1)
- serializers/config.py: Config (1)

Tests:
- 11 backwards compatibility tests added (test_modular_serializers.py)
- All 789 existing tests pass
